feat: ajustes para el primer parcial

+++ src/main/java/com/uniminuto/clinica/repository/MedicoRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
  * @author lmora
  */
 @Repository
-public interface MedicoRepository extends JpaRepository<Medico, Long> {
+public interface MedicoRepository extends JpaRepository<Medico, Integer> {
     List<Medico> findByEspecializacion(Especializacion e);
 }
